+++ www/src/markdown.js
@@ -3,6 +3,7 @@
 const MarkdownIt = require("markdown-it");
 
 const renderer = new MarkdownIt({
+    breaks: true,
     html: false,
     linkify: true,
     typographer: false